requirements.txt
fastapi
uvicorn
spacy
dateparser
python-dotenv
slowapi
